+++ zospy/analyses/psf.py
@@ -28,7 +28,7 @@ def huygens_psf(
 
     Parameters
     ----------
-    oss: zospy.core.OpticStudioSystem
+    oss: zospy.zpcore.OpticStudioSystem
         A ZOSPy OpticStudioSystem instance. Should be sequential.
     pupil_sampling: str | int
         The pupil sampling, either string (e.g. '64x64') or int. The integer will be treated as a ZOSAPI Constants